web/requirements.txt
flask>=3.0
torch>=2.1
torchvision>=0.16
timm>=0.9
opencv-python>=4.8
numpy>=1.24
pillow>=10.0
matplotlib>=3.8
tqdm>=4.60
anthropic>=0.39
